Fix Subscription constructor values in JNI

+++ medialibrary/jni/utils.cpp
@@ -246,7 +246,7 @@ convertSubscriptionObject(JNIEnv* env, fields *fields, medialibrary::Subscriptio
     auto parent = subsPtr->parent();
     jlong parentId = (parent != nullptr) ? (jlong) parent->id() : 0;
     return utils::jni::object{ env, env->NewObject(fields->Subscription.clazz, fields->Subscription.initID,
-            (jlong) subsPtr->id(), (jint) subsPtr->service(), name.get(), parentId, (jint) subsPtr->nbMedia(), (jint) subsPtr->nbUnplayedMedia())
+            (jlong) subsPtr->id(), (jint) subsPtr->service(), name.get(), (jlong) parentId, (jint) subsPtr->nbMedia(), (jint) subsPtr->nbUnplayedMedia())
     };
 }
